Update simfloat for g++ 4.4.1 - aliasing problem
git-svn-id: https://unisim.org/svn/devel@12638 ea533459-3c0a-0410-a024-a90cf4b01ff5

+++ unisim_lib/unisim/util/simfloat/floating.tcc
@@ -3560,7 +3560,7 @@ TBuiltDouble<TypeTraits>::writeDecimal(std::ostream& osOut) const {
          register int uIndexExponent = 0;
          while (thisType(dSource).multAssign(vSuccessiveExponents[uIndexExponent], scfFlags)
                .queryBasicExponent() < TypeTraits::getZeroExponent(biExponent)) {
-            if (uIndexExponent < vSuccessiveExponents.size() - 1) {
+            if (uIndexExponent < (int) vSuccessiveExponents.size() - 1) {
                uExponentAdditional *= 2;
                ++uIndexExponent;
             }
